seperate tests for each class in seperate files
